feat(go): add an opt-in -tags static cgo binding (fully static on musl)
Refactor the binding behind hooks (ensureLoaded/nativeResolve/nativeABIVersion):
binding_purego.go (!static) keeps the default dlopen path, binding_cgo.go (static)
links libsecretspec_ffi.a in via cgo. On Linux the archive is built for musl and
combined with -extldflags -static for a fully-static executable. The default
go get path is unchanged (purego, no cgo, toolchain-free), so static is strictly
additive. stage-staticlib.sh stages the archive + header + generated cgo LDFLAGS.
