+++ secretspec-go/dlopen_windows.go
@@ -1,4 +1,4 @@
-//go:build windows
+//go:build windows && !static
 
 package secretspec
